extract upload input to separate component, improve styles
