+++ src/components/input/index.tsx
@@ -1,1 +1,2 @@
 export * from "./Input";
+export * from "./InputUpload";
